@@ -1,17 +1,5 @@
 module airline_booking_system
 
-go 1.19
+go 1.23
 
-require (
-	github.com/jackc/pgx/v5 v5.3.1
-	github.com/rs/zerolog v1.15.0
-)
-
-require (
-	github.com/jackc/pgpassfile v1.0.0 // indirect
-	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
-	github.com/jackc/puddle/v2 v2.2.0 // indirect
-	golang.org/x/crypto v0.6.0 // indirect
-	golang.org/x/sync v0.1.0 // indirect
-	golang.org/x/text v0.7.0 // indirect
-)
+require github.com/lib/pq v1.10.9
