@@ -2,17 +2,16 @@ package main
 
 import (
 	"context"
-	"fmt"
+	"database/sql"
 	"sync"
+	"sync/atomic"
 	"time"
 
-	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/pgxpool"
-	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
-)
+	"fmt"
+	"log"
 
-const rowCount = 50
+	_ "github.com/lib/pq"
+)
 
 const (
 	host      = "localhost"
@@ -21,144 +20,130 @@ const (
 	password  = "password"
 	dbname    = "default_database"
 	tableName = "seats"
+	seatCount = 1000
 )
 
-func connect(ctx context.Context) (*pgxpool.Pool, error) {
-	connStr := fmt.Sprintf(
-		"host=%s port=%d user=%s password=%s dbname=%s sslmode=disable",
-		host, port, user, password, dbname,
-	)
+func connect(ctx context.Context) *sql.DB {
+	connStr := fmt.Sprintf("host=%s port=%d user=%s password=%s dbname=%s sslmode=disable",
+		host, port, user, password, dbname)
+	db, err := sql.Open("postgres", connStr)
+	if err != nil {
+		log.Fatal(err)
+	}
 
-	cfg, err := pgxpool.ParseConfig(connStr)
-	cfg.MaxConns = 10
-	cfg.MinConns = 2
-	cfg.MaxConnLifetime = 5 * time.Minute
-	conn, err := pgxpool.NewWithConfig(ctx, cfg)
+	err = db.PingContext(ctx)
 	if err != nil {
-		return nil, err
+		log.Fatal(err)
 	}
-	err = conn.Ping(ctx)
+
+	// Configure connection pool
+	db.SetMaxOpenConns(200)
+	db.SetMaxIdleConns(2)
+
+	// Drop the table if it exists
+	_, err = db.ExecContext(ctx, `DROP TABLE IF EXISTS flight_bookings`)
 	if err != nil {
-		return nil, err
+		log.Fatal(err)
 	}
-	return conn, nil
+
+	// Create flight booking table
+	_, err = db.ExecContext(ctx, `CREATE TABLE IF NOT EXISTS flight_bookings (
+	id SERIAL PRIMARY KEY,
+	booked_by VARCHAR(255) UNIQUE
+)`)
+
+	if err != nil {
+		log.Fatal(err)
+	}
+	stmt, err := db.PrepareContext(ctx, "INSERT INTO flight_bookings (booked_by) VALUES ($1)")
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer stmt.Close()
+
+	for range seatCount {
+		_, err = stmt.ExecContext(ctx, nil)
+		if err != nil {
+			log.Fatal(err)
+		}
+	}
+	fmt.Println("Added entries to flight_bookings table.")
+	return db
 }
 
 type Seat struct {
-	Id     int    `db:"id"`
-	Row    int    `db:"row"`
-	Col    string `db:"col"`
-	UserId int    `db:"user_id"`
 }
 
-func show(ctx context.Context, conn *pgxpool.Pool) {
-	for _, col := range []string{"a", "b", "c", "_", "d", "e", "f"} {
-		fmt.Println()
-		if col == "_" {
-			continue
-		}
-		for row := 0; row < rowCount; row++ {
-			qr, err := conn.Query(ctx, "SELECT * FROM seats WHERE col=$1 AND row=$2", col, row)
-			if err != nil {
-				log.Error().Err(err).Msg("failed to fetch row")
-				qr.Close()
-				continue
-			}
-			seats, err := pgx.CollectRows(qr, pgx.RowToStructByName[Seat])
-			if err != nil {
-				log.Error().Err(err).Msg("failed to collect")
-			} else {
-				// fmt.Printf("%03d-", seats[0].UserId)
-				if seats[0].UserId == 0 {
-					fmt.Print(".")
-				} else {
-					fmt.Print("x")
-				}
-			}
-			qr.Close()
-		}
-	}
-	fmt.Println()
+func show(ctx context.Context) {
 }
 
 func main() {
 	ctx := context.Background()
-	zerolog.SetGlobalLevel(zerolog.InfoLevel)
+	db := connect(ctx)
+	defer db.Close()
 
-	conn, err := connect(ctx)
-	if err != nil {
-		log.Fatal().Err(err).Msg("failed to connect")
-	}
+	wg := &sync.WaitGroup{}
+	id := atomic.Int32{}
+	book := func(count int) {
+		defer wg.Done()
+		conn, err := db.Conn(ctx)
+		if err != nil {
+			log.Println("Error getting connection: %v", err)
+			return
+		}
+		defer conn.Close()
 
-	t, err := conn.Exec(ctx, `CREATE TABLE IF NOT EXISTS seats (
-		id SERIAL PRIMARY KEY,
-		row INT NOT NULL,
-		col VARCHAR NOT NULL,
-		user_id INT DEFAULT 0,
-		UNIQUE (row, col)
-	);`)
-	if err != nil {
-		log.Fatal().Err(err).Msg("failed to create table")
-	} else {
-		log.Info().Msg(t.String())
-	}
+		innerWg := &sync.WaitGroup{}
+		defer innerWg.Wait()
 
-	for row := 0; row < rowCount; row++ {
-		for _, col := range []string{"a", "b", "c", "d", "e", "f"} {
-			if t, err := conn.Exec(ctx, "INSERT INTO seats (col, row) VALUES ($1, $2);", col, row); err != nil {
-				log.Error().Err(err).Msg("failed to insert row")
-			} else {
-				log.Debug().Msg(t.String())
-			}
+		for range count {
+			innerWg.Add(1)
+			go func() {
+				defer innerWg.Done()
+				i := id.Add(1)
+				_, err = conn.ExecContext(ctx, `UPDATE flight_bookings
+	SET booked_by = $1
+	WHERE id = (
+		SELECT id
+		FROM flight_bookings
+		WHERE booked_by IS NULL
+		LIMIT 1
+		FOR UPDATE SKIP LOCKED
+	)`, i)
+				if err != nil {
+					log.Printf("Error booking seat: %v", err)
+					return
+				}
+			}()
 		}
 	}
 
-	show(ctx, conn)
-
-	startTime := time.Now()
-
-	wg := &sync.WaitGroup{}
-	wg.Add(6 * rowCount)
-	for u := 1; u <= 6*rowCount; u++ {
-		go func(uid int, wg *sync.WaitGroup) {
-			defer wg.Done()
-			rows, err := conn.Query(ctx, `
-				UPDATE seats
-				SET user_id = $1
-				FROM (
-					SELECT id
-					FROM seats
-					WHERE user_id = 0
-					LIMIT 1
-					FOR UPDATE SKIP LOCKED
-				) AS subquery
-				WHERE seats.id = subquery.id
-				RETURNING seats.id as id, row, col, user_id;`,
-				uid,
-			)
-			if err != nil {
-				log.Error().Err(err).Msg("failed to update")
-				return
-			}
-			seats, err := pgx.CollectRows(rows, pgx.RowToStructByName[Seat])
-			if err != nil {
-				log.Error().Err(err).Msg("failed to collect rows")
-			} else {
-				log.Debug().Interface("seat", seats).Msg("row updated")
-			}
-		}(u, wg)
+	countBookings := func() {
+		var count int
+		err := db.QueryRowContext(ctx, "SELECT COUNT(*) FROM flight_bookings WHERE booked_by IS NOT NULL").Scan(&count)
+		if err != nil {
+			log.Fatal(err)
+		}
+		fmt.Printf("Total bookings: %d\n", count)
+		err = db.QueryRowContext(ctx, "SELECT COUNT(*) FROM flight_bookings WHERE booked_by IS NULL").Scan(&count)
+		if err != nil {
+			log.Fatal(err)
+		}
+		fmt.Printf("Total seats left: %d\n", count)
 	}
 
+	start := time.Now()
+
+	threads := 50
+	for range threads {
+		wg.Add(1)
+		go book(seatCount / threads)
+	}
+
+	fmt.Println("waiting")
 	wg.Wait()
-	endTime := time.Now()
+	fmt.Println("done")
+	fmt.Println(time.Since(start))
 
-	log.Info().
-		Str("delta", endTime.Sub(startTime).String()).
-		Msg("time taken")
-
-	show(ctx, conn)
-
-	if _, err := conn.Exec(ctx, `DROP TABLE seats;`); err != nil {
-		log.Error().Err(err).Msg("failed to drop table")
-	}
+	countBookings()
 }
